% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nontravelers.R
\name{add_nontravelers}
\alias{add_nontravelers}
\title{Add non-travelers to the travelers data set for Mplus models}
\usage{
add_nontravelers(trdat, prdat, usevar)
}
\arguments{
\item{trdat}{Travel-level data frame, which will be modified and used}

\item{prdat}{Person-level data frame to get the full list of person IDs}

\item{usevar}{Variable from \code{trdat} whose unique values will be used to create the analysis variables}
}
\value{
Returns an object that has all people, whether or not they traveled, with the analysis variables ready in columns for Mplus.
}
\description{
This function takes the people who did not travel and adds them to the travel dataset with 0's for all travel variables that will be used in the Mplus analysis.
}
