% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_dirs.R
\name{create_model_dirs}
\alias{create_model_dirs}
\title{Creates Mplus model directory structure}
\usage{
create_model_dirs(model_name)
}
\arguments{
\item{model_name}{A string that will be used to name the analysis folder where all model}
}
\value{
This function creates folders in the right places, or lets you know if the directory structure already exists.
}
\description{
Uses a string to create a directory structure for running Mplus models.
}
