% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLreplication.R
\name{LLrep_to_table}
\alias{LLrep_to_table}
\title{Loglikelihood replication table from \code{mplusModel}}
\usage{
LLrep_to_table(mplusModel)
}
\arguments{
\item{mplusModel}{An \code{mplusModel} object, only of a single model.}
}
\description{
This function takes the output of a single model as an \code{mplusModel} object and returns the table of final stage loglikelihood values.
}
