% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLreplication.R
\name{LLreplication}
\alias{LLreplication}
\title{How many times the maximum loglikelihood was replicated}
\usage{
LLreplication(mplusModel)
}
\arguments{
\item{mplusModel}{An \code{mplusModel} object, only of a single model.}
}
\description{
This function returns a count of the number of times the maximum loglikelihood was replicated
}
