% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dummies.R
\name{make_dummies}
\alias{make_dummies}
\title{Make analysis variables}
\usage{
make_dummies(dat, usevar)
}
\arguments{
\item{dat}{Data frame. The output will return a set of dummies for each observation in this data frame}

\item{usevar}{Variable from \code{dat} whose unique values will be used to create the dummy variables}
}
\value{
Datset with the same observations as is input, but with only dummy variables and personid as output
}
\description{
This function takes a single column and creates a set of dummy columns from it
}
