% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_mplusObj.R
\name{model_to_mplusObjList}
\alias{model_to_mplusObjList}
\title{Extract input from mplus.model into list format suitable for \code{MplusAutomation::mplusObject()}}
\usage{
model_to_mplusObjList(mplusmodelinp)
}
\arguments{
\item{mplusmodelinp}{A single \code{mplus.inp} object}
}
\description{
This function uses \code{tidyverse} functions to extract input from 1 mplus.model so that after running \code{MplusAutomation::mplusObject()} I can use it to modify input using \code{MplusAutomation::update.mplusObject()}
}
