% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mplus_data.R
\name{write_mplus_data}
\alias{write_mplus_data}
\title{Write data file for Mplus, including setting working directory}
\usage{
write_mplus_data(df, wd_for_analysis, filename, writeData, hashfilename)
}
\arguments{
\item{df}{Argument used by \code{MplusAutomation::prepareMplusData}. From the \code{prepareMplusData} documentation: The \code{R} data.frame to be prepared for Mplus.}

\item{wd_for_analysis}{Working directory where you want your .dat file to be saved. This is where your .inp file should also be saved.}

\item{filename}{Argument used by \code{MplusAutomation::prepareMplusData}. From the \code{prepareMplusData} documentation: The filename for the tab-delimited data file for use with Mplus.}

\item{writeData}{Argument used by \code{MplusAutomation::prepareMplusData}. From the \code{prepareMplusData} documentation: A character vector, one of 'always', 'ifmissing', 'never' indicating whether the data files (*.dat) should be written to disk. Defaults to ‘always’ for consistency with previous behavior. See details for further information.}

\item{hashfilename}{Argument used by \code{MplusAutomation::prepareMplusData}. From the \code{prepareMplusData} documentation: A logical whether or not to add a hash of the raw data to the data file name. Defaults to \code{FALSE} for consistency with previous behavior where this feature was not available.}
}
\value{
Writes .dat file formatted for Mplus to the proper working directory. It also returns the Mplus syntax to copy/paste into a .inp file that goes into the same working directory.
}
\description{
This function modifies \code{prepareMplusData} from the \code{MplusAutomation} package. It changes the working directory to allow you to get a shortened file name in the pasted syntax.
}
